miseAjourRomain

+++ database/migrations/2022_05_16_103109_create_comites_table.php
@@ -25,8 +25,8 @@ return new class extends Migration
             $table->string('facebookLink')->nullable();
             $table->string('webSite')->nullable();
             $table->string('description');
-            $table->float('latitude', 8, 6);
-            $table->float('longitude', 8, 6);
+            $table->decimal('latitude', 10, 8);
+            $table->decimal('longitude', 10, 8);
             $table->timestamps();
         });
     }
